Move CommandParserInstance into CommandParser
