+++ src/main/java/com/bortbort/arduino/FiloFirmata/Parser/CommandParser.java
@@ -1,11 +1,16 @@
 package com.bortbort.arduino.FiloFirmata.Parser;
 
+import com.bortbort.arduino.FiloFirmata.FirmataHelper;
 import com.bortbort.arduino.FiloFirmata.Messages.Message;
+import com.bortbort.arduino.FiloFirmata.Parser.Builders.AnalogMessageBuilder;
+import com.bortbort.arduino.FiloFirmata.Parser.Builders.DigitalPortBuilder;
+import com.bortbort.arduino.FiloFirmata.Parser.Builders.ProtocolVersionBuilder;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.InputStream;
 import java.util.HashMap;
+import java.util.Map;
 
 /**
  * Firmata Command Parser.
@@ -14,9 +19,9 @@ import java.util.HashMap;
 public class CommandParser {
     private static final Logger log = LoggerFactory.getLogger(CommandParser.class);
 
-    public static final CommandParserInstance DEFAULT_INSTANCE = new CommandParserInstance(log, new HashMap<>());
+    public static final CommandParser DEFAULT_INSTANCE = new CommandParser(new HashMap<>());
     static {
-        DEFAULT_INSTANCE.addParser(SysexCommandParser.DEFAULT_INSTANCE);
+        DEFAULT_INSTANCE.add(SysexCommandParser.DEFAULT_INSTANCE);
     }
 
     /**
@@ -26,7 +31,7 @@ public class CommandParser {
      * @param messageBuilder MessageBuilder object that translates the byte message into a Message object.
      */
     public static void addParser(MessageBuilder messageBuilder) {
-        DEFAULT_INSTANCE.addParser(messageBuilder);
+        DEFAULT_INSTANCE.add(messageBuilder);
     }
 
     /**
@@ -36,7 +41,7 @@ public class CommandParser {
      */
     public static void addParser(MessageBuilder... messageBuilderList) {
         for (MessageBuilder messageBuilder : messageBuilderList) {
-            DEFAULT_INSTANCE.addParser(messageBuilder);
+            addParser(messageBuilder);
         }
     }
 
@@ -52,11 +57,92 @@ public class CommandParser {
      * @return Message representing the Firmata Message that the SerialPort communications device sent.
      */
     public static Message handleByte(byte commandByte, InputStream inputStream) {
-        return DEFAULT_INSTANCE.handleByte(commandByte, inputStream);
+        return DEFAULT_INSTANCE.handleInputStream(commandByte, inputStream);
     }
 
     /**
-     * CommandParser is a singleton object.
+     * Map of messageBuilders used to parse an array/inputstream of bytes into a Firmata Message object.
      */
-    private CommandParser() {}
+    private final Map<Byte, MessageBuilder> messageBuilderMap;
+
+    public CommandParser() {
+        this(new HashMap<>());
+    }
+
+    public CommandParser(Map<Byte, MessageBuilder> messageBuilderMap) {
+        this.messageBuilderMap = new HashMap<>();
+        addDefaultParser();
+        this.messageBuilderMap.putAll(messageBuilderMap);
+    }
+
+    /**
+     * Register all pre-built Message builder objects available in the Firmata library.
+     */
+    private void addDefaultParser() {
+        add(
+                // Support Sysex commands
+                new SysexCommandParser(),
+                // Support Protocol Version command
+                new ProtocolVersionBuilder(),
+                // Analog command
+                new AnalogMessageBuilder(),
+                // Digital Port command
+                new DigitalPortBuilder()
+        );
+    }
+
+    /**
+     * Add a custom parser to the Firmata library. When the command byte for the parser is received, the parser
+     * will be responsible for turning the the data that follows into a Firmata message.
+     *
+     * @param messageBuilder MessageBuilder object that translates the byte message into a Message object.
+     */
+    public void add(MessageBuilder messageBuilder) {
+        messageBuilderMap.put(messageBuilder.getCommandByte(), messageBuilder);
+    }
+
+    /**
+     * Array input for addParser(MessageBuilder).
+     *
+     * @param messageBuilderList Array of MessageBuilder objects that translate the byte message into a Message object.
+     */
+    public void add(MessageBuilder... messageBuilderList) {
+        for (MessageBuilder messageBuilder : messageBuilderList) {
+            add(messageBuilder);
+        }
+    }
+
+    /**
+     * Attempts to identify the corresponding Firmata Message object that responds to the given commandByte
+     * If found, it will ask for the message to be built using given inputStream. The builder will take away
+     * any bytes necessary to build the message and then return. To attempt to identify the correct message
+     * we must mask the incoming byte against 0xF0 if the byte is less than 0xF0, as commands below 0xF0 are
+     * only legal within this mask, due to potential data padding in the same byte, per the Firmata spec.
+     *
+     * @param commandByte CommandByte representing the identify of a specific command Message packet.
+     * @param inputStream InputStream representing the data following the command byte.
+     * @return Message representing the Firmata Message that the SerialPort communications device sent.
+     */
+    public Message handleInputStream(byte commandByte, InputStream inputStream) {
+        // Firmata command bytes are limited to a 0xF0 mask, or bytes 0xF0-0xFF
+        //   The extra data in the byte is for routing the command to various 'midi channels'
+        //   or in our case Firmata device pins.
+        byte firmataCommandByte = commandByte < (byte) 0xF0 ? (byte) (commandByte & 0xF0) : commandByte;
+        MessageBuilder messageBuilder = messageBuilderMap.get(firmataCommandByte);
+        if (messageBuilder != null) {
+            // Pull out the pin identifier bits that were padded into the byte (if any)
+            byte channelByte = (byte) (commandByte & 0x0F);
+            Message message = messageBuilder.buildMessage(channelByte, inputStream);
+            if (message == null) {
+                log.error("Error building Firmata messageBuilder for command byte {}.",
+                        FirmataHelper.bytesToHexString(commandByte));
+            } else {
+                return message;
+            }
+        } else {
+            log.warn("Dropped byte {}.", FirmataHelper.bytesToHexString(commandByte));
+        }
+
+        return null;
+    }
 }
